Remove now unneeded FingerprintLocation class and use Span instead
FingerprintLocation has same start/end attributes as Span.
