+++ hegpdup/duplicate_finder.py
@@ -4,6 +4,8 @@ import logging
 
 from intervaltree import IntervalTree
 
+from .span import Span
+
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(
@@ -36,24 +38,6 @@ class Duplicate:
 
     def __repr__(self):
         return f"Duplicate(start={self.start}, end={self.end}, fingerprint={self.fingerprint}, fromFingerprint={self.fromFingerprint})"
-
-
-class Span:
-    __slots__ = "start", "end"
-
-    def __init__(self, start, end):
-        self.start = start
-        self.end = end
-
-    @property
-    def length(self):
-        return self.end - self.start
-
-    def __hash__(self):
-        return hash(self.start, self.end)
-
-    def __repr__(self):
-        return f"Span(start={self.start}, end={self.end})"
 
 
 class DuplicateFinder:
